Add integration test for happy path
Test integrated functionality of creating and performing a job
in AWS Elasitc Beanstalk environments.

- Deploys a rails application to an EB web environment
- Deploys same rails application to an EB worker tier environment

The rails application bundles this gem and has the
`active_job.queue_adapater` set to the adapter provided by this gem.
The rails application exposes a REST API which allows
to create, fetch and destroy resources, as well as creating
a job that destroys a specific resource.

The integration test will
- create a resource
- create a job to destroy this resource
- verify if resource has been destroyed within a 5 secondes time window