Do not create queue if it doesn't exist
Raise error if job is bound to a queue that has
not been set up.

This avoids potential confusion, given that a worker environment can not pull
any message from  a queue that did not exist when the worker environment was
launched, because the queue has to be selected when the worker environment is
set up.