Always intercept requests that carry a message digest
Version 1.3.0 introduced a message attributes that identifies
this gem as the origin of SQS messages sent by the
Active Job adapter of this gem (see 16e1a1f8b719bb4b).

However this change breaks backwards compatibility for messages
that have been created by a former version of the adapter.

This commit allows requests which have been created by a former
version to be intercepted.

Refer to #10