Change underscore to dash in message attribute name
The message attribute name will be the HTTP header of that
attribute when the SQS daemon transforms a message into an
HTTP request. If this attribute name contains an underscore,
so will the HTTP header. However, Puma rejects headers
containing underscores in their name, even though they
are valid according to RFC822.

Close #15