Check if remote ip belongs to Docker host
Accept requests from the AWS SQS daemon that are from outside
the docker container if the application is deployed in a
Docker environment and the origin is the Docker host itself.
Docker creates the `docker0` network per default. The application
container and the host will be attached to this network.
The address `172.17.0.1` belongs then to the host in a standard
configuration [1].

Close #26

[1] https://docs.docker.com/engine/userguide/networking/dockernetworks/