Fix #51 Add lazy loading of AWS credential configuration
The default configuration of AWS credentials for this gem
uses AWS instance profiles for retrieving credentials. It
seems that `Aws::InstanceProfileCredentials.new` is blocking
if not executed on an actual EC2 instance, which consequently
impacts the startup time of the application. Therefore,
we postpone the initialization of the credential configuration
to when it's actually needed in the queue adapter.